requirements.txt
requests==2.25.1
pymongo==3.11.2
prettytable==2.2.0
pandas==1.1.3
numpy==1.19.5
pyserial==3.5
# Station==2.0.0
